% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_pressure.R
\name{actual_vapour_pressure}
\alias{actual_vapour_pressure}
\title{Actual Vapour Pressure}
\usage{
actual_vapour_pressure(e_sat, rh)
}
\arguments{
\item{e_sat}{PA or KPA}

\item{rh}{as fraction}
}
\value{
e_act in units of e_sat
}
\description{
Calculated using relative humidity and saturated vapour pressure.
}
\examples{
actual_vapour_pressure(e_sat = 4.360379, rh = 0.35)
}
