% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_pressure.R
\name{actual_vapour_pressure_ah}
\alias{actual_vapour_pressure_ah}
\title{Actual Vapour Pressure (using absolute humidity)}
\usage{
actual_vapour_pressure_ah(absolute_humidity, T_c, R_v = 0.4615)
}
\arguments{
\item{absolute_humidity}{kg/kg or g/g}

\item{T_c}{temperature Celsius}

\item{R_v}{# kPa·K–1·m3·kg–1}
}
\value{
kPa
}
\description{
Another way to get actual vapour pressure as a funciton of absolute humidity. From Stull et al., 2017 eq. 1.19.
}
\examples{
actual_vapour_pressure_ah(absolute_humidity = 0.0148, T_c = 20, R_v = 0.4615)
}
