% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{buckfun}
\alias{buckfun}
\title{Correction of RH when Air Temp < 0 C}
\usage{
buckfun(RH, T_c)
}
\arguments{
\item{RH}{relative humidity (frac or percentage)}

\item{T_c}{air temperature deg C}
}
\value{
adjusted relative humidity value corrected when below deg. C
}
\description{
This is the code used in Harder 2016 for phase correction, also see CRHMr buck function. Reference for this function is Buck 1981
}
\examples{
buckfun(0.5, -10)
}
