% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{clausius_clapeyron}
\alias{clausius_clapeyron}
\title{Saturated Vapour Pressure}
\usage{
clausius_clapeyron(T_c, e_o = 0.6113)
}
\arguments{
\item{T_c}{temperature in Celsius, function does kelvin conversion}

\item{e_o}{vapour pressure constant kpa}
}
\value{
e_sat in kPa
}
\description{
Returns the saturated vapour pressure using the Clausius-Clapeyron equation. From Stull et al., 2017 eq. 4.1
}
\examples{
clausius_clapeyron(T_c = 30, e_o = 0.6113)
}
