% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density_air_virtual}
\alias{density_air_virtual}
\title{Density Air}
\usage{
density_air_virtual(
  T_c,
  p_atm,
  mixing_ratio,
  mixing_ratio_liquid = 0,
  mixing_ratio_ice = 0,
  R_d = 0.287053
)
}
\arguments{
\item{T_c}{temperature Celsius}

\item{p_atm}{atmospheric pressure, kPa}

\item{mixing_ratio}{kg / m3}

\item{mixing_ratio_liquid}{kg / m3 usually assume 0}

\item{mixing_ratio_ice}{kg / m3 usually assume 0}

\item{R_d}{gas constant for unsaturated air kPa·K–1·m3·kg–1}
}
\value{
total density dry + wet g / m3
}
\description{
Density Air using virtual temp, hardly different from above in practice (g / m3). From Stull et al., 2017 eq. 1.23.
}
\examples{
density_dry_air(15, 101.325, 0.03)
}
