% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{density_dry_air}
\alias{density_dry_air}
\title{Density Dry Air}
\usage{
density_dry_air(T_c, p_atm, R_d = 0.287053)
}
\arguments{
\item{T_c}{temperature Celsius}

\item{p_atm}{atmosphereic pressure, kPa}

\item{R_d}{gas constant for unsaturated air kPa·K–1·m3·g–1}
}
\value{
total density of dry air kg / m3
}
\description{
Density Dry Air   From Stull et al., 2017 eq. 1.18.
}
\examples{
density_dry_air(15, 101.325)
}
