% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{density_moist_air}
\alias{density_moist_air}
\title{Density Moist Air}
\usage{
density_moist_air(T_c, p_atm, R_v = 0.4615)
}
\arguments{
\item{T_c}{temperature Celsius}

\item{p_atm}{atmosphereic pressure, kPa}

\item{R_v}{gas constant for moist air kPa·K–1·m3·g–1}
}
\value{
total density of moist air g / m3
}
\description{
Density Dry Air From Stull et al., 2017 eq. 1.19. This one is not recomended in the text because the water vapour constant, is not constant.
}
\examples{
density_moist_air(15, 101.325)
}
