% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{dew_point_temp_e_act}
\alias{dew_point_temp_e_act}
\title{Dew Point Temperature}
\usage{
dew_point_temp_e_act(e_act, R_v_L_v = 0.0001844, e_o = 0.6113)
}
\arguments{
\item{e_act}{unstaturated / actual vapour pressure}

\item{R_v_L_v}{water-vapour gas constant for dry air over the latent heat of vapourization (units are ^ K -1)}

\item{e_o}{0.6113 kPa}
}
\value{
Temperature in Celsius
}
\description{
Calculates dew point temperature using actual vapour pressure. From Stull et al., 2017 eq. 4.15a.
}
\examples{
dew_point_temp_e_act(1.286)
}
