% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{diffusivity_water_vapour}
\alias{diffusivity_water_vapour}
\title{Diffusivity of water vapour in air}
\usage{
diffusivity_water_vapour(T_c)
}
\arguments{
\item{T_c}{air temperature in celcius}
}
\value{
diffusivity of water vapour in air [m2 s-1]
}
\description{
Helper function for ice bulb temp calculation calculates the diffusivity of water vapour in air.
}
\examples{
diffusivity_water_vapour(10)
}
