% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{humidity_ratio_p}
\alias{humidity_ratio_p}
\title{Mixing Ratio (Humidity Ratio)}
\usage{
humidity_ratio_p(e_act, p_atm, E = 622)
}
\arguments{
\item{e_act}{vapour pressure, can be saturated or unsaturated (kPa)}

\item{p_atm}{kPa}

\item{E}{R_d over R_v g/kg is default or g/g with 0.622}
}
\value{
g / kg (saturated or unsaturated depending on e) or g/g if E == 0.622
}
\description{
Calculate the mixing ratio aka humidity ratio using mostly pressure. From Stull et al., 2017 eq. 4.4/4.5. Can use either saturated or unsaturated for e. You get what you put in.
}
\examples{
humidity_ratio_p(e_act = 0.6113, p_atm = 101.325)
}
