% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{ice_bulb_iter}
\alias{ice_bulb_iter}
\title{Ice Bulb Temperature Calculator}
\usage{
ice_bulb_iter(Ta, RH)
}
\arguments{
\item{Ta}{air temperature in deg C, can be vector or single value}

\item{RH}{relative humidity as a fraction, can be vector or single value}
}
\value{
a vector of ice bulb temperatures in deg C
}
\description{
Uses the Newton-Raphston Iteration Functions to converge on an ice bulb temperature. This code was sent from Phil Harder within his phase correction script.
}
\examples{
ice_bulb_iter(-10, 0.5)
}
\author{
Phillip Harder
}
