% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_heat.R
\name{latent_heat_sublimation}
\alias{latent_heat_sublimation}
\title{Calculate the latent heat of sublimation}
\usage{
latent_heat_sublimation(air_temp)
}
\arguments{
\item{air_temp}{Air temperature in deg. C.}
}
\value{
Latent heat of sublimation in Mj kg
}
\description{
From CRHM NewModules_old.cpp line 1249
}
\examples{
latent_heat_vaporisation(0)
}
