% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latent_heat.R
\name{latent_heat_vaporisation}
\alias{latent_heat_vaporisation}
\title{Calculate the latent heat of vaporisation}
\usage{
latent_heat_vaporisation(air_temp)
}
\arguments{
\item{air_temp}{Air temperature in deg. C.}
}
\value{
Latent heat of vaporisation in j kg
}
\description{
From CRHM ClassevapD_Resist.cpp line 258
}
\examples{
latent_heat_vaporisation(0)
}
