% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{mixing_ratio_p}
\alias{mixing_ratio_p}
\title{Mixing Ratio (Humidity Ratio)}
\usage{
mixing_ratio_p(e_act, p_atm, e_o = 0.622)
}
\arguments{
\item{e_act}{vapour pressure, can be saturated or unsaturated (kPa)}

\item{p_atm}{kPa}

\item{e_o}{R_d over R_v (g/g) Vapour pressure constant Stull et al., eq 4.5}
}
\value{
g/g if e_o == 0.622 or g / kg if 622 (saturated or unsaturated depending on e)
}
\description{
Calculate the mixing ratio aka humidity ratio using mostly pressure. From Stull et al., 2017 eq. 4.4/4.5. Can use either saturated or unsaturated for e. You get what you put in.
}
\examples{
mixing_ratio_p(e_act = 0.6113, p_atm = 101.325)
}
