% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{mixing_ratio_t}
\alias{mixing_ratio_t}
\title{Mixing Ratio (Humidity Ratio)}
\usage{
mixing_ratio_t(T_c, T_wb, W_wb)
}
\arguments{
\item{T_c}{temperature Celsius}

\item{T_wb}{wet bulb temperature}

\item{W_wb}{wet bulb humidity ratio}
}
\value{
g / kg
}
\description{
Ratio of mass water vapour to the mass of dry air. calculate hr using mostly temp - from psychometrics chapter 6 ASHRAE. Useful for wetbulb iteration calculation.
}
\examples{
NA
}
