% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{mr_sh}
\alias{mr_sh}
\title{Conversion: specific humidity (mixing ratio) to humidity ratio}
\usage{
mr_sh(mr)
}
\arguments{
\item{mr}{mixing ratio aka specific humidity}
}
\value{
humidity ratio
}
\description{
Conversion: specific humidity (mixing ratio) to humidity ratio
}
\examples{
NA
}
