% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{potential_temp}
\alias{potential_temp}
\title{Potential Temperature}
\usage{
potential_temp(T_c, P)
}
\arguments{
\item{T_c}{Celsius}

\item{P}{kpa}
}
\value{
Temperature Celsius
}
\description{
Potential Temperature
}
\examples{
potential_temp(10, 70)
}
