% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmospheric_pressure.R
\name{pressure_atmosphere}
\alias{pressure_atmosphere}
\title{Non-uniform equation from Stull et al., 2017 eq. 1.17}
\usage{
pressure_atmosphere(m_asl)
}
\arguments{
\item{m_asl}{metres above sea level}
}
\value{
kpa
}
\description{
Non-uniform equation from Stull et al., 2017 eq. 1.17
}
\examples{
pressure_atmosphere(1000)
}
