% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atmospheric_pressure.R
\name{pressure_uniform}
\alias{pressure_uniform}
\title{Pressure Uniform
Return air pressure given height above sea level, assumes uniform temp and pressure. From Stull et al., 2017 eq. 1.9b}
\usage{
pressure_uniform(m_asl)
}
\arguments{
\item{m_asl}{height in metres}
}
\value{
pressure in kPa
}
\description{
Pressure Uniform
Return air pressure given height above sea level, assumes uniform temp and pressure. From Stull et al., 2017 eq. 1.9b
}
\examples{
pressure_uniform(m_asl = 10^4)

}
