% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humidity.R
\name{relative_humidity}
\alias{relative_humidity}
\title{Relative Humidity}
\usage{
relative_humidity(e_act, e_s)
}
\arguments{
\item{e_act}{kPa or PA}

\item{e_s}{same units as e_act}
}
\value{
fraction
}
\description{
Relative Humidity
}
\examples{
relative_humidity(e_act = 1.286, e_s = 2.369)
}
