% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{tetens}
\alias{tetens}
\title{Tetens' formula for saturated vapour pressure}
\usage{
tetens(T_c, e_o = 0.6113)
}
\arguments{
\item{T_c}{temperature Celsius}

\item{e_o}{vapour pressure constant kpa}
}
\value{
kPa or pa if e_o == 610.8 pa
}
\description{
Change e_o to 610.8 for pa. From Stull et al., 2017 eq. 4.2
}
\examples{
tetens(30)
}
