% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{thermal_conductivity_air}
\alias{thermal_conductivity_air}
\title{Thermal conductivity of air}
\usage{
thermal_conductivity_air(Tai)
}
\arguments{
\item{Tai}{}
}
\value{
thermal conductivity of air [J m-1 s-1 K-1]
}
\description{
Helper function for the ice bulb temp calculation that calculates the thermal conductivity of air [J m-1 s-1 K-1]
}
\examples{
thermal_conductivity_air(10)
}
