% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{virtual_temp}
\alias{virtual_temp}
\title{Virtual Temperature}
\usage{
virtual_temp(T_c, mixing_ratio, mixing_ratio_liquid = 0, mixing_ratio_ice = 0)
}
\arguments{
\item{T_c}{air temperature Celsius}

\item{mixing_ratio}{mixing ratio in grams}

\item{mixing_ratio_liquid}{mixing ratio in grams}

\item{mixing_ratio_ice}{mixing ratio in grams}
}
\value{
Temperature in Celsius
}
\description{
Options to include fraction liquid and ice but usually assume to be 0. Stull et al., 2017 eq. 1.22.
}
\examples{
virtual_temp(35, 0.03, mixing_ratio_liquid = 0.01)
}
