% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature.R
\name{wet_bulb_empirical}
\alias{wet_bulb_empirical}
\title{Wet Bulb Temperature}
\usage{
wet_bulb_empirical(T_c, RH = NA, e_act = NA, e_sat = NA)
}
\arguments{
\item{T_c}{temperature Celsius}

\item{RH}{relative humidity fraction}

\item{e_act}{kpa or pa}

\item{e_sat}{same as above}
}
\value{
temperature celsius
}
\description{
Estimated using an empirical function from Stull et al., 2017 eq. 4.19. Need either RH or esat and eact.
}
\examples{
wet_bulb_empirical(20, .5)
}
