% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{wet_bulb_iter}
\alias{wet_bulb_iter}
\title{Wet Bulb Iterator}
\usage{
wet_bulb_iter(T_c, rh, p_atm, iter = 3000)
}
\arguments{
\item{T_c}{air temperature in celsius}

\item{rh}{relative humidity as fraction}

\item{p_atm}{pressure atmosphere (PA)}

\item{iter}{number of iterations}
}
\value{
temperature celsius
}
\description{
Uses a convergence variable to find the wet bulb temp by solving for. Warren ran through this in class.
}
\examples{
wet_bulb_iter(21, 0.37, 101325)
}
