% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapour_pressure.R
\name{clausius_clapeyron}
\alias{clausius_clapeyron}
\title{Clausius-Clapeyron Equation for Saturated Vapour Pressure}
\usage{
clausius_clapeyron(T_c, Lv_Rv = 5423, Ld_Rv = 6139, e_o = 0.6113)
}
\arguments{
\item{T_c}{temperature in Celsius, function does kelvin conversion}

\item{Lv_Rv}{latent heat of vapourization 2.5e6 j kg-1 (for water) over water vapour gas constant 461 J K -1 kg -1}

\item{Ld_Rv}{latent heat of diffusion 2.83e6 J kg-1 (for ice) over water vapour gas constant 461 J K -1 kg -1}

\item{e_o}{vapour pressure constant kpa}
}
\value{
e_sat in kPa
}
\description{
Returns the saturated vapour pressure using the Clausius-Clapeyron equation. From Stull et al., 2017 eq. 4.1
}
\examples{
clausius_clapeyron(T_c = 30, e_o = 0.6113)
}
