% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{clausius_clapeyron_act}
\alias{clausius_clapeyron_act}
\title{Actual Vapour Pressure}
\usage{
clausius_clapeyron_act(T_c, T_d, e_o = 0.6113)
}
\arguments{
\item{T_c}{temperature in Celsius}

\item{T_d}{dew point temperature}

\item{e_o}{vapour pressure constant kpa}
}
\value{
e_act in kPa
}
\description{
Returns the actual or unsaturated vapour pressure using the Clau-sius-Clapeyron equation.
}
\examples{
clausius_clapeyron_act(T_c = 30, T_d = 12.8, e_o = 0.6113)
}
