% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{density_moist_air}
\alias{density_moist_air}
\title{Density Moist Air}
\usage{
density_moist_air(T_c, rh, p_atm, R_v = 0.4615)
}
\arguments{
\item{T_c}{temperature Celsius}

\item{p_atm}{atmosphereic pressure, kPa}

\item{R_v}{gas constant for moist air kPa·K–1·m3·g–1}
}
\value{
density of moist air g / m3
}
\description{
Since total pressure = pressure dry air + pressure water vapour (from lecture 1 slide 6)
}
\examples{
density_moist_air(15, 0, 101.325)
}
