% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-help-txt.R
\docType{package}
\name{psychRomet}
\alias{psychRomet}
\title{psychRomet: Collection of Equations for Psychrometric Calculations}
\description{
This package contains multiple functions used in the course CE826 at the University of Saskachewan. Psychrometrics is concerned with the physical and thermodynamic properties of gas-vapor mixtures.
}
\section{Foo functions}{

The foo functions ...
}

