% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psychrometric.R
\name{specific_humidity}
\alias{specific_humidity}
\title{Specific Humidity}
\usage{
specific_humidity(e_s, total_pressure, vap_pressure = 622)
}
\arguments{
\item{e_s}{saturated vapour pressure kPa}

\item{total_pressure}{total pressure kPa}

\item{vap_pressure}{constant kPa}
}
\value{
g kg-1
}
\description{
From Stull et al., 2017 eq. 4.7
}
\examples{
specific_humidity(e_s = 0.6114, total_pressure = 101.325, vap_pressure = 622)
}
